Refactor Parser and Laid out Rooms
I refactor parser so now it only can take any size room. The parser requires the first line to be empty(signifies parse the whole room) or row of doors. Then it ask for all the block tiles for that room.

In Room folder, I create an abstract room that does the general update(taken from Room itself). Then i created a central of control for rooms so that you can create all the rooms in one place.

Also created the starter room as a layout.