Updated CurrentRoom
I re-implemented changing of location of room in ChangeRoom Method

In DoorColliderLinkCollision I told it to change where link is placed. I dont know if they will make bugs or not.

In switching scene i just change the the currentLocation(which is in currentLocation) to have central point of control in currentRoom